+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/DepthCamera.cpp
@@ -30,9 +30,6 @@ ADepthCamera::ADepthCamera(const FObjectInitializer &ObjectInitializer)
 
 void ADepthCamera::PostPhysTick(UWorld *World, ELevelTick TickType, float DeltaSeconds)
 {
-  if(ReadyToCapture)
-  {
-    FPixelReader::SendPixelsInRenderThread(*this);
-    ReadyToCapture = false;
-  }
+  TRACE_CPUPROFILER_EVENT_SCOPE(ADepthCamera::PostPhysTick);
+  FPixelReader::SendPixelsInRenderThread(*this);
 }