+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/SceneCaptureSensor.h
@@ -282,6 +282,7 @@ public:
 
   /// Blocks until the render thread has finished all it's tasks.
   void WaitForRenderThreadToFinsih() {
+    TRACE_CPUPROFILER_EVENT_SCOPE(ASceneCaptureSensor::WaitForRenderThreadToFinsih);
     FlushRenderingCommands();
   }
 
@@ -320,7 +321,5 @@ protected:
   bool bEnablePostProcessingEffects = true;
 
   FRenderCommandFence RenderFence;
-
-  bool ReadyToCapture = false;
 
 };