+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/SemanticSegmentationCamera.cpp
@@ -26,9 +26,6 @@ ASemanticSegmentationCamera::ASemanticSegmentationCamera(
 
 void ASemanticSegmentationCamera::PostPhysTick(UWorld *World, ELevelTick TickType, float DeltaSeconds)
 {
-  if(ReadyToCapture)
-  {
-    FPixelReader::SendPixelsInRenderThread(*this);
-    ReadyToCapture = false;
-  }
+  TRACE_CPUPROFILER_EVENT_SCOPE(ASemanticSegmentationCamera::PostPhysTick);
+  FPixelReader::SendPixelsInRenderThread(*this);
 }