+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/Sensor.cpp
@@ -24,7 +24,7 @@ void ASensor::BeginPlay()
 {
   Super::BeginPlay();
   OnPostTickDelegate = FWorldDelegates::OnWorldPostActorTick.AddUObject(
-      this, &ASensor::PostPhysTick);
+      this, &ASensor::PostPhysTickInternal);
 }
 
 void ASensor::Set(const FActorDescription &Description)
@@ -40,7 +40,9 @@ void ASensor::Set(const FActorDescription &Description)
 
 void ASensor::Tick(const float DeltaTime)
 {
+  TRACE_CPUPROFILER_EVENT_SCOPE(ASensor::Tick);
   Super::Tick(DeltaTime);
+  ReadyToTick = true;
   PrePhysTick(DeltaTime);
 }
 
@@ -76,4 +78,14 @@ void ASensor::EndPlay(EEndPlayReason::Type EndPlayReason)
   Stream = FDataStream();
 
   FWorldDelegates::OnWorldPostActorTick.Remove(OnPostTickDelegate);
+}
+
+void ASensor::PostPhysTickInternal(UWorld *World, ELevelTick TickType, float DeltaSeconds)
+{
+  TRACE_CPUPROFILER_EVENT_SCOPE(ASensor::PostPhysTickInternal);
+  if(ReadyToTick)
+  {
+    PostPhysTick(World, TickType, DeltaSeconds);
+    ReadyToTick = false;
+  }
 }