+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Vehicle/CarlaWheeledVehicle.h
@@ -15,6 +15,8 @@
 #include "Vehicle/VehiclePhysicsControl.h"
 #include "VehicleVelocityControl.h"
 #include "WheeledVehicleMovementComponent4W.h"
+#include "VehicleAnimInstance.h"
+#include "MovementComponents/BaseCarlaMovementComponent.h"
 
 #include "CoreMinimal.h"
 
@@ -27,6 +29,18 @@
 #include "CarlaWheeledVehicle.generated.h"
 
 class UBoxComponent;
+
+UENUM()
+enum class VehicleWheelLocation : uint8 {
+
+  FL_Wheel = 0,
+  FR_Wheel = 1,
+  BL_Wheel = 2,
+  BR_Wheel = 3,
+  //Use for bikes and bicycles
+  Front_Wheel = 0,
+  Back_Wheel = 1,
+};
 
 /// Base class for CARLA wheeled vehicles.
 UCLASS()
@@ -120,9 +134,14 @@ public:
   FVehiclePhysicsControl GetVehiclePhysicsControl() const;
 
   UFUNCTION(Category = "CARLA Wheeled Vehicle", BlueprintCallable)
+  void RestoreVehiclePhysicsControl();
+
+  UFUNCTION(Category = "CARLA Wheeled Vehicle", BlueprintCallable)
   FVehicleLightState GetVehicleLightState() const;
 
   void ApplyVehiclePhysicsControl(const FVehiclePhysicsControl &PhysicsControl);
+
+  void SetSimulatePhysics(bool enabled);
 
   void SetWheelCollision(UWheeledVehicleMovementComponent4W *Vehicle4W, const FVehiclePhysicsControl &PhysicsControl);
 
@@ -204,6 +223,14 @@ public:
 
   void SetWheelsFrictionScale(TArray<float> &WheelsFrictionScale);
 
+  void SetCarlaMovementComponent(UBaseCarlaMovementComponent* MoementComponent);
+
+  template<typename T = UBaseCarlaMovementComponent>
+  T* GetCarlaMovementComponent() const
+  {
+    return Cast<T>(BaseMovementComponent);
+  }
+
   /// @}
   // ===========================================================================
   /// @name Overriden from AActor
@@ -241,68 +268,33 @@ private:
   InputControl;
 
   FVehicleControl LastAppliedControl;
+  FVehiclePhysicsControl LastPhysicsControl;
 
-
-//-----CARSIM--------------------------------
 public:
 
-  // Enables carsim once enabled it won't turn back to UE4 physics simulation
-  // (for some reason the UE4 physics get meesed up after enabling carsim)
-  UFUNCTION(Category="CARLA Wheeled Vehicle", BlueprintCallable)
-  void EnableCarSim(FString SimfilePath = "");
+  /// Set the rotation of the car wheels indicated by the user
+  /// 0 = FL_VehicleWheel, 1 = FR_VehicleWheel, 2 = BL_VehicleWheel, 3 = BR_VehicleWheel
+  /// An enum could be used, but this is something situational. An uint8 does the trick just fine
+  ///NOTE : This is purely aesthetic. It will not modify the physics of the car in any way
+  UFUNCTION(Category = "CARLA Wheeled Vehicle", BlueprintCallable)
+  void SetWheelSteerDirection(VehicleWheelLocation WheelLocation, float AngleInDeg);
+  
+  UFUNCTION(Category = "CARLA Wheeled Vehicle", BlueprintCallable)
+  float GetWheelSteerAngle(VehicleWheelLocation WheelLocation);
 
-  // Enables usage of carsim terrain
-  UFUNCTION(Category="CARLA Wheeled Vehicle", BlueprintCallable)
-  void UseCarSimRoad(bool bEnabled);
-
-  #ifdef WITH_CARSIM
   virtual FVector GetVelocity() const override;
-  #endif
 
-  UFUNCTION(Category="CARLA Wheeled Vehicle", BlueprintPure)
-  bool IsCarSimEnabled() const;
-
-  virtual void EndPlay(const EEndPlayReason::Type EndPlayReason);
-
-private:
-
-  // On car mesh hit, only works when carsim is enabled
-  UFUNCTION()
-  void OnCarSimHit(AActor *Actor,
-      AActor *OtherActor,
-      FVector NormalImpulse,
-      const FHitResult &Hit);
-
-  // On car mesh overlap, only works when carsim is enabled
-  // (this event triggers when overlapping with static environment)
-  UFUNCTION()
-  void OnCarSimOverlap(UPrimitiveComponent* OverlappedComponent,
-      AActor* OtherActor,
-      UPrimitiveComponent* OtherComp,
-      int32 OtherBodyIndex,
-      bool bFromSweep,
-      const FHitResult & SweepResult);
-
-  UFUNCTION()
-  void SwitchToUE4Physics();
-
-  UFUNCTION()
-  void RevertToCarSimPhysics();
-
-  UPROPERTY(Category="CARLA Wheeled Vehicle", VisibleAnywhere)
-  bool bCarSimEnabled = false;
-
+//-----CARSIM--------------------------------
   UPROPERTY(Category="CARLA Wheeled Vehicle", EditAnywhere)
   float CarSimOriginOffset = 150.f;
+//-------------------------------------------
+private:
+
+  UPROPERTY(Category="CARLA Wheeled Vehicle", VisibleAnywhere)
+  bool bPhysicsEnabled = true;
 
   // Small workarround to allow optional CarSim plugin usage
   UPROPERTY(Category="CARLA Wheeled Vehicle", VisibleAnywhere, BlueprintReadOnly, meta = (AllowPrivateAccess = "true"))
-  UMovementComponent * ExternalMovementComponent;
+  UBaseCarlaMovementComponent * BaseMovementComponent = nullptr;
 
-  #ifdef WITH_CARSIM
-  AActor* OffsetActor;
-  // Casted version of ExternalMovementComponent
-  UCarSimMovementComponent * CarSimMovementComponent;
-  #endif
-  //-------------------------------------------
 };